[438071] Draft to cut off from diagram part.
A new plugin was created (org.eclipse.umlgen.c.modeler.interactions) to
manage all interactions between the reverse/generation processing and
model editing constraints:
- Initialization (structuration) of UML semantic (and potentially
graphical) models.
- Update of potential diagrams.
- Contribution of menus to allows end-user to launch a part of reverse
(with creation of new diagrams) or generation from graphical editors.

The Topcased "template" extension point was migrated (because required
for the reverse processing) and adapted in order to contribute only his
own UML model template and not a graphical one.

A "synchronizer" extension point was defined to contribute his own
behavior to manage these interactions. By default, a simple synchronizer
is contributed to only initialize a UML model from the reverse
processing.
This patch tries to keep the existing architecture as far as possible.
So, at each code location where an "interaction" is required, we
retrieve the contributed synchronizer extension to call the fitting
service. 

A more priority synchronizer could be contributed to initialize any
graphical model in addition to the UML one, to contribute the launch
menus to a specific graphical editor and to manage the right behavior to
update diagrams.

These classes : Template, TemplateDescriptor, TemplatesManager,
AbstractExtensionManager belong to Airbus France. An authorization from
Airbus is required to contribute this patch on Eclipse's Git.


Change-Id: I58d2fb3d047fb52aad0edd6e13ae009df61e6409
Signed-off-by: Cedric Notot <cedric.notot@obeo.fr>
Bug: 438071